I was allocating too little space in decodeUtf8
The problem arises when we have a ByteString of length one (e.g,
"\149") which is invalid UTF-8, and thus replaced by the default
replacement character ("\239\191\189" in UTF-8).

We initially allocate an Array of length one, which we double to
two when we see we can't fit in the replacement character.
However, two is too small since our replacement character is three
bytes long!